dts: make OEM drink their own poison and die
Volume 1.

The dilaceration of a journey ends in a bloody welfare,
unglorified regurgitation of God over no Man's land.
It used to be the land we all lived in, now, it's all death.

I face the fate that of carnage with wisdom,
nor should I face it any other way.
In the resting old days there used to be flowers,
of red color and disgusting perfume.

After all, we gave up on our old land.

Change-Id: I0f10719fb803bbdf25332a9f1301677bfab0ce85
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>